[rf][CD-008] Remove duplicate providers and clean up weather provider implementation
